bin/shtests wrapper: export INSTALLROOT when using other ksh
When specifying another ksh to test using e.g. KSH=/bin/ksh to test
the system's stock ksh, the pty command was not found because it
resides in the arch/*/bin directory. The main shtests script bases
its $PATH on $INSTALLROOT, but does not set INSTALLROOT itself if
the shell to test is outside the installation hierarchy. Setting it
would allow the pty.sh regresssion tests to run on a shell
specified on the command line.

bin/shtests:
- Set and export INSTALLROOT to our default installation hierarchy
  if it wasn't already set on entry.