emacs.c: unbreak menu-driven pathname completion (re: e8b3274a)
One of the few AT&T fixes applied in early 2020 was a one-line
change to emacs.c tab handling, with only this info in the commit
message:

|  - fix to emacs.c (I think from dgk)

So, it's unknown what that was meant to accomplish, but I did just
find that it breaks menu-driven pathname completion:

$ ls arch/darwin.i386-64/
1) bin/
2) fun/
3) include/
4) lib/
5) man/
6) src/
$ ls arch/darwin.i386-64/3	_

Typing 3+TAB should have inserted 'include/' but inserted a literal
tab instead. Reverting the vague "fix" fixes this bug.