data/signals.c: fix empty SIGINT/SIGPIPE messages
src/cmd/ksh93/data/signals.c includes two checks for the JOBS
identifier; if it is not defined then the interactive shell's
background job signal messages for SIGINT and SIGPIPE are empty.
The cause was that the "jobs.h" header, which defines that ID, was
not #included in signals.c. This commit adds that #include.
(ksh 93u+, ksh 93v- and ksh2020 all have this bug as well.)

Before:

$ sleep 30 &
[1]	86430
$ kill -s INT "$!"
[1] +                          sleep 30 &
$

After:

$ sleep 30 &
[1]	86445
$ kill -s INT "$!"
[1] + Interrupt                sleep 30 &
$